+++ library/src/lib/combobox-input/combobox-input.component.ts
@@ -34,6 +34,10 @@ export class ComboboxInputComponent extends SearchInputComponent {
     @Output()
     newItemClicked: EventEmitter<void> = new EventEmitter<void>();
 
+    /** Text to use for the '+ New Item' option. */
+    @Input()
+    newItemText: string = '+ New Item';
+
     /** @hidden */
     newItemKeydownHandler(event) {
         if (event.code === 'Enter') {